data_pipeline/requirements.txt
openai
pandas
openpyxl
requests
beautifulsoup4
tiktoken
python-dotenv
fastapi
uvicorn
apscheduler